src/requirements.txt
python-dotenv 
elasticsearch
datasets
sentence_transformers
bert-extractive-summarizer